refactor into JsonSearch service
